full_conv/Exp2_Red_Green/requirements_red_green.txt
python=3.7.4
numpy=1.17.2
pytorch>=1.2.0
